updated server code, does it work? no. will it? maybe.
